+++ exporter/exporterhelper/logshelper.go
@@ -45,8 +45,12 @@ func newLogsRequest(ctx context.Context, ld pdata.Logs, pusher PushLogs) request
 	}
 }
 
-func (req *logsRequest) onPartialError(partialErr consumererror.PartialError) request {
-	return newLogsRequest(req.ctx, partialErr.GetLogs(), req.pusher)
+func (req *logsRequest) onError(err error) request {
+	var logError consumererror.Logs
+	if consumererror.AsLogs(err, &logError) {
+		return newLogsRequest(req.ctx, logError.GetLogs(), req.pusher)
+	}
+	return req
 }
 
 func (req *logsRequest) export(ctx context.Context) error {